% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alphabet.R
\name{alphabet}
\alias{alphabet}
\title{Generates Random Letters}
\usage{
alphabet(x)
}
\arguments{
\item{x}{specifies how many letters to return.}
}
\value{
returns the letters
}
\description{
Generates Random Letters
}
